Quality's Delight craft-down integration
Actually, this is a fix between Quality's Delight and Quality Crops, NOT for Lucrative itself.

In Quality Crops, you can craft down a quality crop into it's original.
In Quality's Delight, you CANNOT do this. So, I made it possible.

This has been tested and proven to work in datapack form.